# Conflicts:
#	.idea/workspace.xml
#	frontend/hrm/src/App.js
#	src/main/java/com/humanresourcemanagementsystem/Config/SecurityConfig.java
#	src/main/java/com/humanresourcemanagementsystem/Dto/EmployeeDTO.java
#	src/main/java/com/humanresourcemanagementsystem/Entity/Employee.java
#	src/main/java/com/humanresourcemanagementsystem/Repo/DepartmentRepository.java
#	src/main/java/com/humanresourcemanagementsystem/Service/DepartmentService.java
#	src/main/java/com/humanresourcemanagementsystem/Service/Impl/EmployeeIMPL.java
#	src/main/resources/application.properties

+++ src/main/java/com/humanresourcemanagementsystem/Service/DepartmentService.java
@@ -20,19 +20,16 @@ public class DepartmentService {
     @Autowired
     private DepartmentRepository departmentRepository;
 
-    //Display all Departments
     public List<DepartmentDTO> getAllDepartments() {
         return departmentRepository.findAll().stream()
                 .map(this::convertToDTO)
                 .collect(Collectors.toList());
     }
 
-    //Display department by ID
-    public DepartmentDTO getDepartmentById(Long id) {
+    public DepartmentDTO getDepartmentById(int id) {
         return departmentRepository.findById(id).map(this::convertToDTO).orElse(null);
     }
 
-    //Create Department
     public ResponseData createDepartment(DepartmentDTO departmentDTO) {
         Optional<Department> existingDepartment = departmentRepository.findByDepartmentName(departmentDTO.getDepartmentName());
 
@@ -55,8 +52,7 @@ public class DepartmentService {
         return new ResponseData("Department created successfully.", true, responseData);
     }
 
-    //Update department by ID
-    public DepartmentDTO updateDepartment(Long id, DepartmentDTO departmentDTO) {
+    public DepartmentDTO updateDepartment(int id, DepartmentDTO departmentDTO) {
         if (departmentRepository.existsById(id)) {
             Department department = departmentRepository.findById(id).orElseThrow();
             department.setDepartmentName(departmentDTO.getDepartmentName());
@@ -68,8 +64,7 @@ public class DepartmentService {
         return null;
     }
 
-    //Delete department by ID
-    public String deleteDepartment(Long id) {
+    public String deleteDepartment(int id) {
         Optional<Department> departmentOptional = departmentRepository.findById(id);
 
         if (departmentOptional.isPresent()) {
